Changed the way to get_element path
First the element paths were hardcoded in the code.
Now made a new file Elements_Search_Path.py that holds all the paths in variables.

It's easy to edit and change path variables and also can be used in multiple places without repeating.

Also added benifit that others can see the path all at one place and not cluster the code. Makes the code more readable.

It's not the best/preferred way, but it gets work done soo it's okay